Split up the XAudio implementation into multiple files, for readability
