+++ dlls/xaudio2_7/xaudio_private.h
@@ -148,5 +148,35 @@ extern const IXAudio22Vtbl XAudio22_Vtbl DECLSPEC_HIDDEN;
 extern const IXAudio27Vtbl XAudio27_Vtbl DECLSPEC_HIDDEN;
 #endif
 
+/* xapofx.c */
 extern HRESULT make_xapo_factory(REFCLSID clsid, REFIID riid, void **ppv) DECLSPEC_HIDDEN;
+
+/* xaudio_engine.c */
+extern HRESULT make_xaudio2_factory(REFIID riid, void **ppv) DECLSPEC_HIDDEN;
 extern HRESULT xaudio2_initialize(IXAudio2Impl *This, UINT32 flags, XAUDIO2_PROCESSOR proc) DECLSPEC_HIDDEN;
+
+/* xaudio_voice_source.c */
+extern XA2VoiceImpl *impl_from_IXAudio2SourceVoice(IXAudio2SourceVoice *iface) DECLSPEC_HIDDEN;
+extern const IXAudio2SourceVoiceVtbl XAudio2SourceVoice_Vtbl DECLSPEC_HIDDEN;
+
+/* xaudio_voice_submix.c */
+extern XA2VoiceImpl *impl_from_IXAudio2SubmixVoice(IXAudio2SubmixVoice *iface) DECLSPEC_HIDDEN;
+extern const struct IXAudio2SubmixVoiceVtbl XAudio2SubmixVoice_Vtbl DECLSPEC_HIDDEN;
+
+/* xaudio_voice_master.c */
+extern XA2VoiceImpl *impl_from_IXAudio2MasteringVoice(IXAudio2MasteringVoice *iface) DECLSPEC_HIDDEN;
+extern const struct IXAudio2MasteringVoiceVtbl XAudio2MasteringVoice_Vtbl DECLSPEC_HIDDEN;
+
+/* xaudio_voice_wrap.c */
+extern FAudioVoiceSends *wrap_voice_sends(const XAUDIO2_VOICE_SENDS *sends) DECLSPEC_HIDDEN;
+extern void free_voice_sends(FAudioVoiceSends *sends) DECLSPEC_HIDDEN;
+extern void destroy_voice(XA2VoiceImpl *This) DECLSPEC_HIDDEN;
+extern XA2VoiceImpl *impl_from_IXAudio2Voice(IXAudio2Voice *iface) DECLSPEC_HIDDEN;
+
+/* xaudio_callbacks.c */
+extern const FAudioVoiceCallback FAudioVoiceCallback_Vtbl DECLSPEC_HIDDEN;
+extern const FAudioEngineCallback FAudioEngineCallback_Vtbl DECLSPEC_HIDDEN;
+
+/* xapo_wrap.c */
+extern FAudioEffectChain *wrap_effect_chain(const XAUDIO2_EFFECT_CHAIN *pEffectChain) DECLSPEC_HIDDEN;
+extern void free_effect_chain(FAudioEffectChain *chain) DECLSPEC_HIDDEN;
