+++ dlls/xaudio2_9/Makefile.in
@@ -9,6 +9,13 @@ C_SRCS = \
 	xapofx.c \
 	x3daudio.c \
 	xaudio_dll.c \
+	xaudio_engine.c \
+	xaudio_voice_source.c \
+	xaudio_voice_submix.c \
+	xaudio_voice_master.c \
+	xaudio_voice_wrap.c \
+	xaudio_callbacks.c \
+	xapo_wrap.c \
 	FAudio/FAPOBase.c \
 	FAudio/FAPOFX.c \
 	FAudio/FAPOFX_eq.c \
